requirements.txt
fastapi
openai
faiss-cpu
sentence-transformers
PyPDF2
